fix: 툴팁 모달 뒤에 있는 카드에도 뜨는 문제 수정
